FIX : 로그인 성공시 유저정보 반환

+++ src/main/java/com/nawabali/nawabali/dto/UserDto.java
@@ -31,8 +31,8 @@ public class UserDto {
         String city;
         String district;
         UserRankEnum rank;
-        Long totalLikesCount;
-        Long totalLocalLikesCount;
+        int totalLikesCount;
+        int totalLocalLikesCount;
 
 
         public UserInfoResponseDto(User user) {
